requirements.txt
flask>=2.0
flask-cors>=3.0
python-dotenv>=1.0
supabase>=2.0
psycopg2-binary>=2.9
google-api-python-client>=2.0
google-auth>=2.0
